pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adaptive-spectral-temporal-gnn-for-molecular-toxicity-prediction"
version = "0.1.0"
description = "Molecular toxicity prediction using adaptive spectral-temporal GNN with curriculum learning"
authors = [{name = "Alireza Shojaei"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torch-geometric>=2.3.0",
    "dgl>=1.1.0",
    "networkx>=3.0",
    "deepchem>=2.7.0",
    "tensorflow>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.2.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "mlflow>=2.8.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "rdkit>=2023.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
